pyproject.toml
[project]
name = "erir-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic-settings>=2.10.1",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
erir-scraper = "erir_scraper:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
